Massive refactor: give graph analysis its own package, and split the different flow scanners into individual classes
